Probe PowerHAL HIDL first
Right now AIDL doesn't provide an API for vendor code to query if the
AIDL HAL declared. This will lead to 1) long blocking call for device
doesn't have the AIDL HAL and also 2) potential problem if the service
starts late after 5s timeout.

This is a workaround for addressing 1) on devices having HIDL, but still
affects device that doesn't have HIDL service.

Bug: 149797408
Test: build
Change-Id: I60aada1ab1a044db965987bca5f6eee5b6262b20
Signed-off-by: Albert I <kras@raphielgang.org>